chore: add Ip CRUD to workspace.
