+++ backend/src/controllers/v2/workspaceController.ts
@@ -1,3 +1,4 @@
+import { MongoError } from 'mongodb';
 import { Request, Response } from 'express';
 import * as Sentry from '@sentry/node';
 import { Types } from 'mongoose';
@@ -21,6 +22,10 @@ import {
 import { pushKeys } from '../../helpers/key';
 import { TelemetryService, EventService } from '../../services';
 import { eventPushSecrets } from '../../events';
+import to from 'await-to-js';
+import { Builder } from 'builder-pattern';
+import IPAddress, { IIPAddress } from '../../models/IPAddress';
+import { BadRequestError } from '../../utils/errors';
 
 interface V2PushSecret {
 	type: string; // personal or shared
@@ -507,3 +512,43 @@ export const toggleAutoCapitalization = async (req: Request, res: Response) => {
 		workspace
 	});
 };
+
+
+export const createWorkspaceIPAddress = async (req: Request, res: Response) => {
+  const { workspaceId } = req.params;
+  const { ip } = req.body;
+  const ipAddressToCreate = Builder<IIPAddress>()
+      .ip(ip)
+      .workspace(new Types.ObjectId(workspaceId))
+      .user(new Types.ObjectId(req.user._id))
+      .build();
+
+  const [err, createIPAddress] = await to(IPAddress.create(ipAddressToCreate));
+  if (err) {
+    if ((err as MongoError).code === 11000) {
+      throw BadRequestError({ message: "Ip must be unique in a workspace" });
+    }
+    throw err;
+  }
+  return res.status(201).json(createIPAddress);
+};
+
+export const deleteWorkSpaceIPAddress = async (req: Request, res: Response) => {
+  const { ipId, workspaceId } = req.params;
+  const ipAddressFromDB = await IPAddress.findOne({
+    id: new Types.ObjectId(ipId),
+    workspace: new Types.ObjectId(workspaceId),
+  })
+  if (!ipAddressFromDB) {
+    throw BadRequestError({message: "id does not exist"});
+  }
+  await ipAddressFromDB.deleteOne();
+  return res.json();
+};
+
+export const getWorkspaceIPAddress = async (req: Request, res: Response) => {
+  const { workspaceId } = req.params;
+  const ips = await IPAddress.find({ workspace: new Types.ObjectId(workspaceId) }, ["ip"]);
+  return res.json(ips);
+};
+
